+++ worker/test/src/RTC/TestRtpStreamSend.cpp
@@ -2,6 +2,7 @@
 #include "RTC/Codecs/AV1.hpp"
 #include "RTC/Codecs/PayloadDescriptorHandler.hpp"
 #include "RTC/Codecs/VP8.hpp"
+#include "RTC/Consts.hpp"
 #include "RTC/RTCP/FeedbackRtpNack.hpp"
 #include "RTC/RtpPacket.hpp"
 #include "RTC/RtpStream.hpp"
@@ -900,6 +901,58 @@ SCENARIO("NACK and RTP packets retransmission", "[rtp][rtcp][nack][rtpstreamsend
 		stream->ReceiveNack(&nackPacket2);
 
 		REQUIRE(testRtpStreamListener.retransmittedPackets.empty());
+	}
+
+	SECTION("duplicated packets are discarded")
+	{
+		auto packet(CreateRtpPacket(rtpBuffer1, sizeof(rtpBuffer1), 50001, 1000001));
+
+		// Create a RtpStreamSend instance.
+		TestRtpStreamListener testRtpStreamListener;
+
+		RtpStream::Params params;
+
+		params.ssrc          = packet->GetSsrc();
+		params.clockRate     = 90000;
+		params.useNack       = true;
+		params.mimeType.type = RTC::RtpCodecMimeType::Type::VIDEO;
+
+		std::string mid;
+		auto stream = std::make_unique<RtpStreamSend>(&testRtpStreamListener, params, mid);
+
+		RTC::SharedRtpPacket sharedPacket;
+
+		auto result = stream->ReceivePacket(packet.get(), sharedPacket);
+
+		REQUIRE(result == RTC::RtpStreamSend::ReceivePacketResult::ACCEPTED_AND_STORED);
+
+		result = stream->ReceivePacket(packet.get(), sharedPacket);
+
+		REQUIRE(result == RTC::RtpStreamSend::ReceivePacketResult::DISCARDED);
+	}
+
+	SECTION("packets larger than MTU size are discarded")
+	{
+		auto packet(CreateRtpPacket(rtpBuffer1, RTC::Consts::MtuSize + 1, 50001, 1000001));
+
+		// Create a RtpStreamSend instance.
+		TestRtpStreamListener testRtpStreamListener;
+
+		RtpStream::Params params;
+
+		params.ssrc          = packet->GetSsrc();
+		params.clockRate     = 90000;
+		params.useNack       = true;
+		params.mimeType.type = RTC::RtpCodecMimeType::Type::VIDEO;
+
+		std::string mid;
+		auto stream = std::make_unique<RtpStreamSend>(&testRtpStreamListener, params, mid);
+
+		RTC::SharedRtpPacket sharedPacket;
+
+		auto result = stream->ReceivePacket(packet.get(), sharedPacket);
+
+		REQUIRE(result == RTC::RtpStreamSend::ReceivePacketResult::DISCARDED);
 	}
 
 #ifdef PERFORMANCE_TEST